move, update: FakeRepositoryを移動し、偽予報にも対応
